osd: Add config option to compact OSD's object store on start
By setting this configuration option an OSD will compact it's
store's OMAP on start.

This way admin's can trigger an offline compaction by setting this
configuration value to 'true' and then restarting the OSD.

This is easier than using tools like 'ceph-kvstore-tool' with requires
more manual work on the CLI and might be more difficult for users.

Signed-off-by: Wido den Hollander <wido@42on.com>